Added new TileMap to define types of floor.
The new TileMap uses the name of each atlas to define the name of the
SFX to play when the Player is touching it.
Player can play different FS based on the type of floor she is touching.
Added a new Area2D to the Player so it handles the behavior related to
the types of floor.
Ready to export a new tilemap to be used for painting floors.